fix: fixed a bug to show tag filter result correctly
